switched to async
this commit adds some traits, moves to static dispatch and other stuff

more refactoring

more refactoring
