+++ crates/trainer-tests/Cargo.toml
@@ -11,6 +11,8 @@ repository.workspace = true
 [dependencies]
 api = {path = "../api"}
 sqlite = {path = "../sqlite"}
+tokio = {workspace = true}
+async-trait = {workspace = true}
 
 [dev-dependencies]
 tempfile = "3.14.0"
