+++ crates/trainer-tests/src/lib.rs
@@ -1,2 +1,1 @@
-#[warn(unused_imports)]
-mod exercise_manager_tests;
+
